adjust tests flow
